+++ src/main/java/net/meander/subtlyd/mixin/client/OptionsMixin.java
@@ -1,6 +1,6 @@
 package net.meander.subtlyd.mixin.client;
 
-import net.meander.subtlyd.client.OptionsSD;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -9,12 +9,21 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Options.class)
 public class OptionsMixin {
-    @Inject(method = "processOptions", at = @At("HEAD"))
-    private void saveOptions(Options.FieldAccess access, CallbackInfo ci) {
-        access.process("camera_shake", OptionsSD.cameraShake());
-        access.process("experimental_gui", OptionsSD.gui());
-        access.process("shield_crouch", OptionsSD.shieldCrouch());
-        access.process("advanced_entity_animations", OptionsSD.advancedEntityAnimations());
-        access.process("shield_animation", OptionsSD.shieldAnimation());
+    @Inject(method = "processOptions", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Options;processDumpedOptions(Lnet/minecraft/client/Options$OptionAccess;)V"))
+    private void processOptions(Options.FieldAccess access, CallbackInfo ci) {
+        Options options = Minecraft.getInstance().options;
+
+        access.process("camera_shake", options.cameraShake());
+        access.process("experimental_gui", options.experimentalGui());
+        access.process("shield_crouch", options.shieldCrouch());
+        access.process("fancy_entities", options.fancyEntities());
+        access.process("shield_animation", options.shieldAnimation());
+    }
+
+    @Inject(method = "processDumpedOptions(Lnet/minecraft/client/Options$OptionAccess;)V", at = @At("HEAD"))
+    private void processDumpedOptions(Options.OptionAccess access, CallbackInfo ci) {
+        Options options = Minecraft.getInstance().options;
+
+        access.process("entity_culling", options.entityCulling());
     }
 }