+++ src/main/java/net/meander/subtlyd/mixin/client/gui/screens/options/AccessibilityOptionsScreenMixin.java
@@ -1,6 +1,5 @@
 package net.meander.subtlyd.mixin.client.gui.screens.options;
 
-import net.meander.subtlyd.client.OptionsSD;
 import net.minecraft.client.OptionInstance;
 import net.minecraft.client.Options;
 import net.minecraft.client.gui.screens.options.AccessibilityOptionsScreen;
@@ -18,8 +17,8 @@ public class AccessibilityOptionsScreenMixin {
     private static void options(Options options, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
         List<OptionInstance<?>> optionInstanceList = new ArrayList<>(List.of(cir.getReturnValue().clone()));
 
-        optionInstanceList.add(18, OptionsSD.cameraShake());
-        optionInstanceList.add(22, OptionsSD.shieldAnimation());
+        optionInstanceList.add(12, options.cameraShake());
+        optionInstanceList.add(24, options.shieldAnimation());
         cir.setReturnValue(optionInstanceList.toArray(new OptionInstance[0]));
     }
 }