+++ src/main/java/net/meander/subtlyd/mixin/client/gui/screens/options/VideoSettingsScreenMixin.java
@@ -1,6 +1,5 @@
 package net.meander.subtlyd.mixin.client.gui.screens.options;
 
-import net.meander.subtlyd.client.OptionsSD;
 import net.minecraft.client.OptionInstance;
 import net.minecraft.client.Options;
 import net.minecraft.client.gui.screens.options.VideoSettingsScreen;
@@ -17,7 +16,7 @@ public class VideoSettingsScreenMixin { @Inject(method = "displayOptions", at = 
     private static void displayOptions(Options options, boolean supportsExclusiveFullscreen, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
         List<OptionInstance<?>> optionInstanceList = new ArrayList<>(List.of(cir.getReturnValue().clone()));
 
-        optionInstanceList.add(8, OptionsSD.gui());
+        optionInstanceList.add(8, options.experimentalGui());
         cir.setReturnValue(optionInstanceList.toArray(new OptionInstance[0]));
     }
 
@@ -25,7 +24,16 @@ public class VideoSettingsScreenMixin { @Inject(method = "displayOptions", at = 
     private static void qualityOptions(Options options, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
         List<OptionInstance<?>> optionInstanceList = new ArrayList<>(List.of(cir.getReturnValue().clone()));
 
-        optionInstanceList.add(10, OptionsSD.advancedEntityAnimations());
+        optionInstanceList.add(9, options.fancyEntities());
+        optionInstanceList.add(11, options.entityCulling());
+        cir.setReturnValue(optionInstanceList.toArray(new OptionInstance[0]));
+    }
+
+    @Inject(method = "preferenceOptions", at = @At("RETURN"), cancellable = true)
+    private static void preferenceOptions(Options options, CallbackInfoReturnable<OptionInstance<?>[]> cir) {
+        List<OptionInstance<?>> optionInstanceList = new ArrayList<>(List.of(cir.getReturnValue().clone()));
+
+        optionInstanceList.add(4, options.experimentalGui());
         cir.setReturnValue(optionInstanceList.toArray(new OptionInstance[0]));
     }
 }