+++ src/main/java/net/meander/subtlyd/mixin/client/level/CameraMixin.java
@@ -1,13 +1,14 @@
 package net.meander.subtlyd.mixin.client.level;
 
-import net.meander.subtlyd.client.OptionsSD;
 import net.meander.subtlyd.client.camera.shake.CameraShake;
 import net.minecraft.client.Camera;
 import net.minecraft.client.DeltaTracker;
+import net.minecraft.client.Minecraft;
 import net.minecraft.util.Mth;
 import net.minecraft.util.Util;
 import net.minecraft.world.entity.Entity;
 import org.jspecify.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -16,6 +17,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(Camera.class)
 public abstract class CameraMixin {
+    @Shadow @Final private Minecraft minecraft;
+
     @Shadow protected abstract void setRotation(float yRot, float xRot);
     @Shadow public abstract @Nullable Entity entity();
     @Shadow public abstract float xRot();
@@ -30,7 +33,7 @@ public abstract class CameraMixin {
      * Determines if camera shake is enabled and if so, handles the math for placing the camera to create the camera shake effect.
      */
     private void applyCameraShake() {
-        if (OptionsSD.cameraShake().get()) {
+        if (minecraft.options.cameraShake().get()) {
             float intensity = CameraShake.getShakeIntensity() * 0.5F;
 
             if (intensity > 0.0F) {