+++ src/main/java/net/meander/subtlyd/mixin/client/renderer/item/ItemInHandRendererMixin.java
@@ -1,8 +1,8 @@
 package net.meander.subtlyd.mixin.client.renderer.item;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import net.meander.subtlyd.client.OptionsSD;
 import net.meander.subtlyd.tags.ItemTagsSD;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.ItemInHandRenderer;
 import net.minecraft.client.renderer.SubmitNodeCollector;
@@ -18,7 +18,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(ItemInHandRenderer.class)
 public class ItemInHandRendererMixin {
     @Inject(method = "submitArmWithItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/ItemInHandRenderer;renderItem(Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/item/ItemDisplayContext;Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V"))
-    private void renderBowShake(AbstractClientPlayer player, float frameInterp, float xRot, InteractionHand hand, float attack, ItemStack itemStack, float inverseArmHeight, PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, CallbackInfo ci) {
+    private void renderBowShake(AbstractClientPlayer player, float frameInterp, float xRot, InteractionHand hand, float swingAnimation, ItemStack itemStack, float inverseArmHeight, PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, CallbackInfo ci) {
         if (player.isUsingItem()) {
             if (player.getUsedItemHand() == hand && itemStack.is(Items.BOW)) {
                 float timeHeld = itemStack.getUseDuration(player) - (player.getUseItemRemainingTicks() - frameInterp + 1.0F);
@@ -34,8 +34,8 @@ public class ItemInHandRendererMixin {
     }
 
     @Inject(method = "submitArmWithItem", at = @At("HEAD"), cancellable = true)
-    private void hideItem(AbstractClientPlayer player, float frameInterp, float xRot, InteractionHand hand, float attack, ItemStack itemStack, float inverseArmHeight, PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, CallbackInfo ci) {
-        if (!OptionsSD.shieldAnimation().get()) {
+    private void hideItem(AbstractClientPlayer player, float frameInterp, float xRot, InteractionHand hand, float swingAnimation, ItemStack itemStack, float inverseArmHeight, PoseStack poseStack, SubmitNodeCollector submitNodeCollector, int lightCoords, CallbackInfo ci) {
+        if (!Minecraft.getInstance().options.shieldAnimation().get()) {
             if (player.isUsingItem()) {
                 if (!player.isScoping()) {
                     if (player.getUsedItemHand() == hand && itemStack.is(ItemTagsSD.SHIELDS)) {